requirements.txt
asyncio>=3.4.3
pyserial>=3.5
pyyaml>=6.0
